requirements.txt
# 核心框架 - 2025年最新版本
Django==5.2.*
djangorestframework==3.15.*

# 认证和安全 - 最新版本
djangorestframework-simplejwt==5.5.*
django-cors-headers==4.7.0

# API文档 - 最新版本
drf-spectacular==0.28.0

# 数据库 - 最新版本
mysqlclient==2.2.7

# HTTP客户端 - 现代化选择
httpx==0.28.*

# 开发工具 - 最新版本
django-debug-toolbar==5.2.*
django-extensions==3.2.*

# 测试 - 最新版本
pytest-django==4.8.*
factory-boy==3.3.*

# 代码质量 - 最新版本
black==24.*
isort==5.13.*
mypy==1.8.*
ruff==0.7.*

# 生产环境
gunicorn==23.*
python-dotenv==1.0.*
